+++ token-client/src/main/java/com/sap/cloud/security/xsuaa/Assertions.java
@@ -1,6 +1,6 @@
 package com.sap.cloud.security.xsuaa;
 
-public class ObjectsUtil {
+public class Assertions {
 	public static void assertNotNull(Object object, String message) {
 		if (object == null) {
 			throw new IllegalArgumentException(message);
